+++ src/app/shared/_component/header-page/header-page.component.html
@@ -1,14 +1,14 @@
 <div class="title-container">
-    <label-icon [icon]="titleIcon" [iconOutlined]="iconOutlined" [size]="48"></label-icon>
-    <div class="title">
-        <div *ngIf="title || templateTitle" class="main">
-          <span *ngIf="title">{{title}}</span>
-          <ng-container *ngIf="templateTitle" [ngTemplateOutlet]="templateTitle"></ng-container>
-        </div>
-        <div *ngIf="subTitle || templateSubTitle" class="sub">
-            <span *ngIf="subTitle">{{subTitle}}</span>
-            <ng-container *ngIf="templateSubTitle" [ngTemplateOutlet]="templateSubTitle"></ng-container>
-        </div>
+  <mat-icon  [ngClass]="{ 'material-symbols-outlined': iconOutlined }" class="title-icon">{{ titleIcon }}</mat-icon>
+  <div class="title-content">
+    <h1 *ngIf="title || templateTitle" class="title-label">
+      <span *ngIf="title">{{title}}</span>
+      <ng-container *ngIf="templateTitle" [ngTemplateOutlet]="templateTitle"></ng-container>
+    </h1>
+    <div *ngIf="subTitle || templateSubTitle" class="subtitle-label">
+      <span *ngIf="subTitle">{{subTitle}}</span>
+      <ng-container *ngIf="templateSubTitle" [ngTemplateOutlet]="templateSubTitle"></ng-container>
     </div>
+  </div>
 </div>
 <ng-content></ng-content>
