+++ src/app/views/deploiment/deploiment.component.scss
@@ -1,188 +1,141 @@
-// ===========================
-// En-tête de Page Moderne
-// ===========================
-.header-container {
-  padding: 14px 16px;
+.stats-container {
   display: flex;
   align-items: center;
-  justify-content: space-between;
-  gap: 32px;
-  flex-wrap: wrap;
+  gap: 24px;
 
-  .title-container {
-    display: flex;
-    align-items: center;
-    gap: 16px;
-
-    .title-icon {
-      font-size: 40px;
-      width: 40px;
-      height: 40px;
-      color: #3b82f6;
-      flex-shrink: 0;
-    }
-
-    .title-content {
-      display: flex;
-      flex-direction: column;
-      gap: 4px;
-
-      .title-label {
-        font-size: 24px;
-        font-weight: 700;
-        color: #1e293b;
-        margin: 0;
-        line-height: 1.2;
-      }
-
-      .subtitle-label {
-        font-size: 14px;
-        color: #64748b;
-        margin: 0;
-        font-weight: 500;
-      }
-    }
+  .stat-divider {
+    width: 1px;
+    height: 48px;
+    background: #e2e8f0;
   }
 
-  .stats-container {
+  .stat-item {
     display: flex;
     align-items: center;
-    gap: 24px;
+    gap: 12px;
+    padding: 12px 16px;
+    border-radius: 10px;
 
-    .stat-divider {
-      width: 1px;
-      height: 48px;
-      background: #e2e8f0;
-    }
-
-    .stat-item {
+    .stat-icon-wrapper {
+      width: 40px;
+      height: 40px;
+      border-radius: 10px;
       display: flex;
       align-items: center;
-      gap: 12px;
-      padding: 12px 16px;
-      border-radius: 10px;
+      justify-content: center;
+      flex-shrink: 0;
+
+      mat-icon {
+        font-size: 22px;
+        width: 22px;
+        height: 22px;
+      }
+    }
+
+    .stat-details {
+      display: flex;
+      flex-direction: column;
+      gap: 2px;
+
+      .stat-value {
+        font-size: 20px;
+        font-weight: 700;
+        line-height: 1;
+        font-variant-numeric: tabular-nums;
+      }
+
+      .stat-label {
+        font-size: 11px;
+        font-weight: 600;
+        text-transform: uppercase;
+        letter-spacing: 0.05em;
+      }
+    }
+
+    &.stat-online {
+      background: rgba(16, 185, 129, 0.08);
 
       .stat-icon-wrapper {
-        width: 40px;
-        height: 40px;
-        border-radius: 10px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        flex-shrink: 0;
+        background: rgba(16, 185, 129, 0.15);
 
         mat-icon {
-          font-size: 22px;
-          width: 22px;
-          height: 22px;
+          color: #10b981;
         }
       }
 
       .stat-details {
-        display: flex;
-        flex-direction: column;
-        gap: 2px;
-
         .stat-value {
-          font-size: 20px;
-          font-weight: 700;
-          line-height: 1;
-          font-variant-numeric: tabular-nums;
+          color: #059669;
         }
 
         .stat-label {
-          font-size: 11px;
-          font-weight: 600;
-          text-transform: uppercase;
-          letter-spacing: 0.05em;
+          color: #047857;
+        }
+      }
+    }
+
+    &.stat-pending {
+      background: rgba(245, 158, 11, 0.08);
+
+      .stat-icon-wrapper {
+        background: rgba(245, 158, 11, 0.15);
+
+        mat-icon {
+          color: #f59e0b;
         }
       }
 
-      &.stat-online {
-        background: rgba(16, 185, 129, 0.08);
-
-        .stat-icon-wrapper {
-          background: rgba(16, 185, 129, 0.15);
-
-          mat-icon {
-            color: #10b981;
-          }
+      .stat-details {
+        .stat-value {
+          color: #d97706;
         }
 
-        .stat-details {
-          .stat-value {
-            color: #059669;
-          }
+        .stat-label {
+          color: #b45309;
+        }
+      }
+    }
 
-          .stat-label {
-            color: #047857;
-          }
+    &.stat-offline {
+      background: rgba(100, 116, 139, 0.08);
+
+      .stat-icon-wrapper {
+        background: rgba(100, 116, 139, 0.15);
+
+        mat-icon {
+          color: #64748b;
         }
       }
 
-      &.stat-pending {
-        background: rgba(245, 158, 11, 0.08);
-
-        .stat-icon-wrapper {
-          background: rgba(245, 158, 11, 0.15);
-
-          mat-icon {
-            color: #f59e0b;
-          }
+      .stat-details {
+        .stat-value {
+          color: #475569;
         }
 
-        .stat-details {
-          .stat-value {
-            color: #d97706;
-          }
+        .stat-label {
+          color: #64748b;
+        }
+      }
+    }
 
-          .stat-label {
-            color: #b45309;
-          }
+    &.stat-total {
+      background: rgba(59, 130, 246, 0.08);
+
+      .stat-icon-wrapper {
+        background: rgba(59, 130, 246, 0.15);
+
+        mat-icon {
+          color: #3b82f6;
         }
       }
 
-      &.stat-offline {
-        background: rgba(100, 116, 139, 0.08);
-
-        .stat-icon-wrapper {
-          background: rgba(100, 116, 139, 0.15);
-
-          mat-icon {
-            color: #64748b;
-          }
+      .stat-details {
+        .stat-value {
+          color: #2563eb;
         }
 
-        .stat-details {
-          .stat-value {
-            color: #475569;
-          }
-
-          .stat-label {
-            color: #64748b;
-          }
-        }
-      }
-
-      &.stat-total {
-        background: rgba(59, 130, 246, 0.08);
-
-        .stat-icon-wrapper {
-          background: rgba(59, 130, 246, 0.15);
-
-          mat-icon {
-            color: #3b82f6;
-          }
-        }
-
-        .stat-details {
-          .stat-value {
-            color: #2563eb;
-          }
-
-          .stat-label {
-            color: #1d4ed8;
-          }
+        .stat-label {
+          color: #1d4ed8;
         }
       }
     }
