+++ src/app/views/supervision/_component/client/client-supervision.view.html
@@ -53,7 +53,6 @@
       <mat-icon matSuffix>search</mat-icon>
     </button>
   </div>
-
 </header-page>
 <div style="display: flex;">
   <div style="width: 50%" class="card-container mat-elevation-z5">
