+++ src/app/views/supervision/_component/server/server-supervision.view.html
@@ -55,6 +55,7 @@
   </div>
 
 </header-page>
+
 <div style="display: flex;">
   <div style="width: 50%" class="card-container mat-elevation-z5">
     <div class="title-container">
