+++ src/app/views/supervision/_component/server/server-supervision.view.ts
@@ -9,7 +9,6 @@ import {ChartProvider, field} from "@oneteme/jquery-core";
 import {InstanceTraceService} from "../../../../service/jquery/instance-trace.service";
 import {DatePipe, DecimalPipe, Location} from "@angular/common";
 import {MatDialog} from "@angular/material/dialog";
-import {StacktraceDialogComponent} from "../stacktrace-dialog/stacktrace-dialog.component";
 import {DateAdapter, MAT_DATE_FORMATS} from "@angular/material/core";
 import {CustomDateAdapter} from "../../../../shared/material/custom-date-adapter";
 import {MY_DATE_FORMATS} from "../../../../shared/shared.module";
@@ -20,6 +19,9 @@ import {ConfigDialogComponent} from "../config-dialog/config-dialog.component";
 import {InstanceService} from "../../../../service/jquery/instance.service";
 import {ServerInstanceSelectorDialogComponent} from "./server-instance-selector-dialog/server-instance-selector-dialog.component";
 import {MatSnackBar} from "@angular/material/snack-bar";
+import {
+  StacktraceDialogComponent
+} from "../../../../shared/_component/exception-display/stacktrace-dialog/stacktrace-dialog.component";
 
 @Component({
   templateUrl: './server-supervision.view.html',
@@ -508,7 +510,7 @@ export class ServerSupervisionView implements OnInit, OnDestroy {
 
   open(row: any) {
     this._dialog.open(StacktraceDialogComponent, {
-      data: row
+      data: { message: row.message, stackTraceRows: row.stacktrace }
     });
   }
 
@@ -540,12 +542,6 @@ export class ServerSupervisionView implements OnInit, OnDestroy {
     this.formGroup.patchValue({
       server: server
     }, { emitEvent: false });
-  }
-
-  openConfig() {
-    this._dialog.open(ConfigDialogComponent, {
-      data: this.instance.configuration
-    });
   }
 
   openInstanceSelector() {
