--- src/app/views/supervision/_component/stacktrace-dialog/stacktrace-dialog.component.scss
@@ -1,4 +1,0 @@
-.stacktrace-textarea {
-  max-height: 300px;
-  overflow-y: auto;
-}
