+++ src/app/views/views.module.ts
@@ -63,7 +63,6 @@ import {DetailTimelineComponent} from "./detail/session/_component/timeline/deta
 import {ActionTableComponent} from "./detail/session/_component/action-table/action-table.component";
 import {DetailRequestView} from "./detail/request/detail-request.view";
 import {DetailRestView} from "./detail/request/rest/detail-rest.view";
-import {StacktraceDialogComponent} from "./supervision/_component/stacktrace-dialog/stacktrace-dialog.component";
 import {ConfigDialogComponent} from "./supervision/_component/config-dialog/config-dialog.component";
 import {ReportTableComponent} from "./supervision/_component/report-table/report-table.component";
 import {DetailStageTableComponent} from "./detail/session/_component/stage-table/detail-stage-table.component";
@@ -138,7 +137,6 @@ import {DetailLogTableComponent} from "./detail/session/_component/log-table/det
     AnalyticView,
     ActionTableComponent,
     SearchRequestView,
-    StacktraceDialogComponent,
     ConfigDialogComponent,
     ReportTableComponent,
     DetailStageTableComponent,
